mailing/views.py
from django.shortcuts import render
from rest_framework import viewsets
from .models import Mailing, Message, Client
from .serializers import MailingSerializer, MessageSerializer, ClientSerializer


class MailingViewSet(viewsets.ModelViewSet):
    serializer_class = MailingSerializer
    queryset = Mailing.objects.all()

class MessageViewSet(viewsets.ModelViewSet):
    serializer_class = MessageSerializer
    queryset = Message.objects.all()

class ClientViewSet(viewsets.ModelViewSet):
    serializer_class = ClientSerializer
    queryset = Client.objects.all()


